Add messaging and annotation to video cal activity

+++ app/src/main/java/com/revis/ui/message/Position.kt
@@ -3,6 +3,6 @@ package com.revis.ui.message
 import com.google.gson.annotations.SerializedName
 
 data class Position(
-    @SerializedName("x") val x: Int,
-    @SerializedName("y") val y: Int
+    @SerializedName("x") val x: Float,
+    @SerializedName("y") val y: Float
 )
